+++ streamlit/db_setup.py
@@ -32,9 +32,9 @@ import psycopg2
 import psycopg2.extras
 from dotenv import load_dotenv
 
-from utils import log_message
-from db_scripts import execute_single_query, execute_multi_query
-from constants import ErrorMessages
+from utils.utils import log_message
+from utils.db_scripts import execute_single_query, execute_multi_query
+from utils.constants import ErrorMessages
 
 
 load_dotenv()
