+++ streamlit/pages/0_⬆️_Upload_Content.py
@@ -12,9 +12,9 @@ import uuid
 import pandas as pd
 import streamlit as st
 
-from utils import log_message
-from formatting import convert_to_json
-from db_scripts import execute_multi_query
+from utils.utils import log_message
+from utils.formatting import convert_to_json
+from utils.db_scripts import execute_multi_query
 
 
 # Set page configuration
