+++ streamlit/pages/6_🏋️‍♀️_Improve & Evaluate.py
@@ -4,10 +4,9 @@ import json
 from dotenv import load_dotenv
 import plotly.express as px
 import re
-import openai
-import os
-from utils import  log_message,clear_all_caches, calculate_success_failure_rate
-from db_scripts import (
+
+from utils.utils import  log_message,clear_all_caches, calculate_success_failure_rate
+from utils.db_scripts import (
     fetch_bad_lesson_plans, fetch_result_data, 
     fetch_lesson_plan_json, insert_single_lesson_plan, 
     get_prompt, fetch_prompt_objectives_desc,  
@@ -16,9 +15,10 @@ from db_scripts import (
     start_experiment, 
     fetch_final_data, 
     delete_created_sample, delete_lesson_plans_from_sample_lesson_plans
-    
 )
-from inference import run_agent_openai_inference
+
+from utils.inference import run_agent_openai_inference
+
 import json
 import matplotlib.pyplot as plt
 import warnings
