+++ streamlit/utils/db_scripts.py
@@ -56,7 +56,6 @@
     Deletes lesson plans from the sample_lesson_plans table.
     """
 
-from utils import *
 import uuid
 import hashlib
 import psycopg2
@@ -66,7 +65,7 @@ import pandas as pd
 import streamlit as st
 from formatting import fix_json_format
 from constants import ErrorMessages
-from utils import log_message
+from utils import log_message, get_env_variable
 
 def get_db_connection():
     """ Establish a connection to the PostgreSQL database.
@@ -884,3 +883,29 @@ def delete_lesson_plans_from_sample_lesson_plans(sample_id):
     except Exception as e:
         print(f"An error occurred while deleting the sample lesson plans: {e}")
         return False
+    
+
+
+def get_lesson_plans_for_dataset(keyword=None):
+    """ Retrieve lesson plans from the lesson_plans table based on a 
+        keyword filter.
+
+    Args:
+        keyword (str, optional): Keyword to filter generation details. 
+            Defaults to None.
+
+    Returns:
+        pd.DataFrame: DataFrame containing lesson plan IDs and 
+            generation details.
+    """
+    query = """
+        SELECT lp.id, lp.generation_details
+        FROM lesson_plans lp
+        WHERE 1=1
+    """
+    params = []
+    if keyword:
+        query += " AND lp.generation_details LIKE %s"
+        params.append(f"%{keyword}%")
+
+    return execute_single_query(query, params, return_dataframe=True)
